requirements.txt
streamlit==1.37.0
pandas==2.2.2
openpyxl==3.1.2
plotly==5.17.0
pyyaml==6.0.1
ldap3